[opencl] rms_norm: take epsilon as a float, drop the 1e-4 fp16 floor
The fp16 RMSNorm kernels took epsilon as a `half`, and the GPU rms_norm
layer floored it to 1e-4 so that it survived that argument type, on the
assumption that this is negligible for a normal row. It is not for a model
whose embedding rows are small. With embedding rows whose mean square is
near 2e-4 (rms ~0.014), eps=1e-4 in place of the configured 1e-5 scales
the first attention norm by up to 17%: the relative error of that norm's
output against a reference implementation was 0.17, on the OpenCL lanes
only. CUDA and the CPU path use the configured value.

The consequence was visible in the text, not only in the residuals: on a
1024-token teacher-forced prompt one greedy step chose the token the
reference ranks 38th (logit gap 10.5), on an Intel Xe iGPU and on an Adreno
alike. With the configured epsilon the same norm's relative error is 3e-4
and every remaining divergence from the reference is a rank-2 near tie.

All five kernels in rmsnorm_fp16.cl and the fused norm+add kernel already
reduce in float, so the argument simply becomes a float and the host binds
the float it was given; a zero row still yields 0 * rsqrt(eps) = 0, which
is what the floor was guarding. 1e-5 and 1e-6 are subnormal as halves, so
the conversion was also at the mercy of the device's denormal handling.

Behaviour change: OpenCL output changes for every model whose rms_norm_eps
is below 1e-4 (the gemma family configures 1e-6), and moves towards what
the CUDA and CPU lanes already produce. The kernel signature and all of its
host binders change in this one commit. The on-disk kernel binary cache is
keyed on the kernel source, so it rebuilds by itself.

+++ nntrainer/tensor/cl_operations/cl_kernels/rmsnorm_fp16.cl
@@ -19,7 +19,7 @@ __kernel void
 rmsnorm_cl_fp16(__global const half *input, // Input tensor
                 __global half *output,      // Output tensor
                 __global const half *alpha, // Alpha values (one for each width)
-                half epsilon,
+                float epsilon,
                 int B, // Number of batches
                 int C, // Number of channels
                 int H, // Height of feature map
@@ -66,7 +66,7 @@ RMSN_SG_ATTR
 __attribute__((reqd_work_group_size(RMSN_LWS, 1, 1)))
 __kernel void rmsnorm_cl_fp16_coop(__global const half *input,
                                    __global half *output,
-                                   __global const half *alpha, half epsilon,
+                                   __global const half *alpha, float epsilon,
                                    int n_rows, int W) {
   const int row = get_group_id(0);
   const int tid = get_local_id(0);
@@ -132,7 +132,7 @@ __kernel void rmsnorm_cl_fp16_coop(__global const half *input,
 __attribute__((reqd_work_group_size(RMSN_LWS, 1, 1)))
 __kernel void rmsnorm_cl_fp16_coop_ng(__global const half *input,
                                       __global half *output,
-                                      __global const half *alpha, half epsilon,
+                                      __global const half *alpha, float epsilon,
                                       int n_rows, int W) {
   const int row = get_group_id(0);
   const int tid = get_local_id(0);
@@ -180,7 +180,7 @@ __attribute__((reqd_work_group_size(RMSN_LWS, 1, 1)))
 __kernel void rms_reverse_norm_cl_fp16_coop(__global const half *input,
                                             __global half *output,
                                             __global const half *weight,
-                                            half out_scale, half epsilon,
+                                            half out_scale, float epsilon,
                                             int n_rows, int W) {
   const int row = get_group_id(0);
   const int tid = get_local_id(0);
@@ -244,7 +244,7 @@ __kernel void rms_reverse_norm_cl_fp16_coop(__global const half *input,
 #define RMSNQ_LWS_MAX 256
 __kernel void rmsnorm_cl_fp16_coop_q(__global const half *input,
                                      __global half *output,
-                                     __global const half *alpha, half epsilon,
+                                     __global const half *alpha, float epsilon,
                                      int n_rows, int W,
                                      __global char *act_int8,
                                      __global float *scale_per_row,
